+++ src/App-Services/Add-VNet-integration-to-AppService.ps1
@@ -8,7 +8,7 @@ param (
     [Parameter(Mandatory)][string] $AppServiceVnetIntegrationSubnetName,
     [Parameter()][string] $AppServiceSlotName,
     [Parameter()][bool] $ApplyToAllSlots = $false,
-    [Parameter()][bool] $RouteAllTrafficThroughVnet = $true
+    [Parameter()][bool] $RouteAllTrafficThroughVnet = $false
 )
 
 #region ===BEGIN IMPORTS===
@@ -23,18 +23,21 @@ if ($ApplyToAllSlots -eq $True)
     $availableSlots = Invoke-Executable -AllowToFail az webapp deployment slot list --name $AppServiceName --resource-group $AppServiceResourceGroupName | ConvertFrom-Json
 }
 
+#fetching the Identifier of the Vnet and subnet
+$vnetsubnetIdentifier = Get-VnetSubnetIdentifiers -VnetName $AppServiceVnetIntegrationVnetName -SubnetName $AppServiceVnetIntegrationSubnetName
+
 if (![string]::IsNullOrWhiteSpace($AppServiceVnetIntegrationVnetResourceGroupName))
 {
     $AppServiceVnetIntegrationVnetName = (Invoke-Executable az network vnet show --resource-group $AppServiceVnetIntegrationVnetResourceGroupName --name $AppServiceVnetIntegrationVnetName | ConvertFrom-Json).id
 }
 
 # Set VNET Integration on the main given slot (normally production)
-Add-VnetIntegration -appType 'webapp' -AppResourceGroupName $AppServiceResourceGroupName -AppName $AppServiceName -AppVnetIntegrationVnetName $AppServiceVnetIntegrationVnetName -AppVnetIntegrationSubnetName $AppServiceVnetIntegrationSubnetName -AppSlotName $AppServiceSlotName -RouteAllTrafficThroughVnet:$RouteAllTrafficThroughVnet
+Add-VnetIntegration -appType 'webapp' -AppResourceGroupName $AppServiceResourceGroupName -AppName $AppServiceName -AppVnetIntegrationVnetIdentifier $vnetsubnetIdentifier.VnetIdentifier -AppVnetIntegrationSubnetIdentifier $vnetsubnetIdentifier.SubnetIdentifier -AppSlotName $AppServiceSlotName -RouteAllTrafficThroughVnet:$RouteAllTrafficThroughVnet
 
 # Apply to all slots if desired
 foreach ($availableSlot in $availableSlots)
 {
-    Add-VnetIntegration -appType 'webapp' -AppResourceGroupName $AppServiceResourceGroupName -AppName $AppServiceName -AppVnetIntegrationVnetName $AppServiceVnetIntegrationVnetName -AppVnetIntegrationSubnetName $AppServiceVnetIntegrationSubnetName -AppSlotName $availableSlot.name -RouteAllTrafficThroughVnet:$RouteAllTrafficThroughVnet
+    Add-VnetIntegration -appType 'webapp' -AppResourceGroupName $AppServiceResourceGroupName -AppName $AppServiceName -AppVnetIntegrationVnetIdentifier $vnetsubnetIdentifier.VnetIdentifier -AppVnetIntegrationSubnetIdentifier $vnetsubnetIdentifier.SubnetIdentifier -AppSlotName $availableSlot.name -RouteAllTrafficThroughVnet:$RouteAllTrafficThroughVnet
 }
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet