+++ src/AzDocs.Common/public/DeploymentSlots-Helper-Functions.ps1
@@ -6,13 +6,11 @@ function New-DeploymentSlot
         [Parameter(Mandatory)][string] $ResourceResourceGroupName,
         [Parameter(Mandatory)][string] $ResourceName,
         [Parameter(Mandatory)][string] $ResourceDeploymentSlotName,
-        [Parameter(Mandatory)][string] $ResourceAppServicePlanTier,
-        [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
         [Parameter()][bool] $StopResourceSlotImmediatelyAfterCreation,
         [Parameter()][System.Object[]] $ResourceTags,
         [Parameter()][string] $ResourceNumberOfInstances = 2,
         [Parameter()][bool] $ResourceAlwaysOn = $true,
-        [Parameter()][string][ValidateSet("1.0", "1.1", "1.2")] $ResourceMinimalTlsVersion = "1.2",
+        [Parameter()][string][ValidateSet('1.0', '1.1', '1.2')] $ResourceMinimalTlsVersion = '1.2',
         [Parameter()][bool] $DisablePublicAccessForResourceDeploymentSlot = $true,
         # Forcefully agree to this resource to be spun up to be publicly available
         [Parameter()][switch] $ForcePublic,
@@ -30,7 +28,17 @@ function New-DeploymentSlot
         [Parameter()][string] $ResourcePrivateEndpointVnetName,
         [Parameter()][string] $ResourcePrivateEndpointSubnetName,
         [Parameter()][string] $DNSZoneResourceGroupName,
-        [Parameter()][string] $ResourcePrivateDnsZoneName
+        [Parameter()][string] $ResourcePrivateDnsZoneName,
+
+        # Diagnostic settings
+        [Parameter()][System.Object[]] $DiagnosticSettingsLogs,
+        [Parameter()][System.Object[]] $DiagnosticSettingsMetrics,
+        [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
+
+        # Disable diagnostic settings
+        [Parameter()][switch] $DiagnosticSettingsDisabled,
+        
+        [Parameter()][string[]] $CORSUrls = @()
     )
 
     Write-Header -ScopedPSCmdlet $PSCmdlet
@@ -56,15 +64,34 @@ function New-DeploymentSlot
     Invoke-Executable az $AppType config set --ids $resourceSlotId --number-of-workers $ResourceNumberOfInstances --always-on $ResourceAlwaysOn --ftps-state Disabled --min-tls-version $ResourceMinimalTlsVersion --slot $ResourceDeploymentSlotName
     Invoke-Executable az $Apptype identity assign --ids $resourceSlotId --slot $ResourceDeploymentSlotName
     
-    Set-ConfigurationForResource -AppType $AppType -ResourceSlotId $resourceSlotId -ResourceResourceGroupName $ResourceResourceGroupName -ResourceDeploymentSlotName $ResourceDeploymentSlotName -ResourceAppServicePlanTier $ResourceAppServicePlanTier -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -ResourceName $ResourceName
+    Set-ConfigurationForResource -AppType $AppType -ResourceSlotId $resourceSlotId -ResourceResourceGroupName $ResourceResourceGroupName -ResourceDeploymentSlotName $ResourceDeploymentSlotName -ResourceName $ResourceName
+
+    # Set cors settings if they're provided, please note that CORS settings do not get swapped
+    if ($CorsUrls)
+    {
+        Set-CorsSettings -AppType $AppType -CORSUrls $CorsUrls -ResourceId $resourceSlotId -ResourceDeploymentSlotName $ResourceDeploymentSlotName
+    }
+
+    # Set Diagnostic Settings
+    if ($DiagnosticSettingsDisabled)
+    {
+        Remove-DiagnosticSetting -ResourceId $resourceSlotId -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -ResourceName $ResourceName
+    }
+    else
+    {
+        Set-DiagnosticSettings -ResourceId $resourceSlotId -ResourceName $ResourceName -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -DiagnosticSettingsLogs:$DiagnosticSettingsLogs -DiagnosticSettingsMetrics:$DiagnosticSettingsMetrics 
+    }
 
     # Update Tags
-    Set-ResourceTagsForResource -ResourceId $resourceSlotId -ResourceTags ${ResourceTags}
+    if ($ResourceTags)
+    {
+        Set-ResourceTagsForResource -ResourceId $resourceSlotId -ResourceTags ${ResourceTags}
+    }
 
     # VNET Whitelisting
     if (!$ResourceDisableVNetWhitelisting -and $GatewayVnetResourceGroupName -and $GatewayVnetName -and $GatewaySubnetName)
     {
-        Write-Host "VNET Whitelisting is desired. Adding the needed components."
+        Write-Host 'VNET Whitelisting is desired. Adding the needed components.'
         $RootPath = (Get-Item $PSScriptRoot).Parent.Parent
         switch ($AppType)
         {
@@ -77,7 +104,7 @@ function New-DeploymentSlot
     if (!$ResourceDisablePrivateEndpoints -and $ResourcePrivateEndpointVnetResourceGroupName -and $ResourcePrivateEndpointVnetName -and $ResourcePrivateEndpointSubnetName -and $DNSZoneResourceGroupName -and $ResourcePrivateDnsZoneName)
     {
         # Fetch needed information
-        Write-Host "A private endpoint is desired. Adding the needed components."
+        Write-Host 'A private endpoint is desired. Adding the needed components.'
         $vnetId = (Invoke-Executable az network vnet show --resource-group $ResourcePrivateEndpointVnetResourceGroupName --name $ResourcePrivateEndpointVnetName | ConvertFrom-Json).id
         $resourcePrivateEndpointSubnetId = (Invoke-Executable az network vnet subnet show --resource-group $ResourcePrivateEndpointVnetResourceGroupName --name $ResourcePrivateEndpointSubnetName --vnet-name $ResourcePrivateEndpointVnetName | ConvertFrom-Json).id
         $resourcePrivateEndpointName = Get-PrivateEndpointNameForResource -AppType $AppType -ResourceName $ResourceName -ResourceDeploymentSlotName $ResourceDeploymentSlotName
@@ -90,7 +117,7 @@ function New-DeploymentSlot
     # Disable public acces, if it's not public
     if (!$ForcePublic -and $DisablePublicAccessForResourceDeploymentSlot)
     {
-        Write-Host "Disabling public access on the deployment slot."
+        Write-Host 'Disabling public access on the deployment slot.'
 
         $accessRestrictionRuleName = 'DisablePublicAccess'
         $cidr = '0.0.0.0/0'
@@ -136,8 +163,6 @@ function Set-ConfigurationForResource
         [Parameter(Mandatory)][string] $ResourceSlotId,
         [Parameter(Mandatory)][string] $ResourceResourceGroupName,
         [Parameter(Mandatory)][string] $ResourceDeploymentSlotName,
-        [Parameter(Mandatory)][string] $ResourceAppServicePlanTier,
-        [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
         [Parameter(Mandatory)][string] $ResourceName
     )
     
@@ -150,20 +175,85 @@ function Set-ConfigurationForResource
             # Enforce HTTPS
             Invoke-Executable az webapp update --resource-group $ResourceResourceGroupName --name $ResourceName --slot $ResourceDeploymentSlotName --https-only true
             Invoke-Executable az webapp log config --ids $ResourceSlotId --detailed-error-messages true --docker-container-logging filesystem --failed-request-tracing true --level warning --web-server-logging filesystem --slot $ResourceDeploymentSlotName
-        
-            $diagnosticSettingsForWebapp = Get-DiagnosticSettingBasedOnTier -ResourceType $AppType -CurrentResourceTier $ResourceAppServicePlanTier
-            if ($diagnosticSettingsForWebapp.DiagnosticSettingType -eq 'Logs')
-            {
-                Set-DiagnosticSettings -ResourceId $ResourceSlotId -ResourceName $ResourceName -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -Logs $diagnosticSettingsForWebapp.DiagnosticSettingValue -Metrics "[ { 'category': 'AllMetrics', 'enabled': true } ]".Replace("'", '\"')
-            }
         }
         'functionapp'
         {
             # Enforce HTTPS
             Invoke-Executable az functionapp update --ids $ResourceSlotId --set httpsOnly=true
-
             Invoke-Executable az functionapp config appsettings set --ids $ResourceSlotId --settings "ASPNETCORE_ENVIRONMENT=$($ASPNETCORE_ENVIRONMENT)" "FUNCTIONS_EXTENSION_VERSION=$($FUNCTIONS_EXTENSION_VERSION)"
-            Set-DiagnosticSettings -ResourceId $resourceSlotId -ResourceName $ResourceName -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -Logs "[{ 'category': 'FunctionAppLogs', 'enabled': true } ]".Replace("'", '\"') -Metrics "[ { 'category': 'AllMetrics', 'enabled': true } ]".Replace("'", '\"')
+        }
+    }
+
+    Write-Footer -ScopedPSCmdlet $PSCmdlet
+}
+
+function Get-DefaultCorsSettings
+{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)][string] [ValidateSet('functionapp', 'webapp')]$AppType
+    )
+
+    Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    $CORSUrls = @()
+    switch ($AppType)
+    {
+        # At this moment, there are no default cors settings that need to be set for webapps. Built in the functionality for future use.
+        'functionapp'
+        {
+            $CORSUrls += 'https://functions.azure.com'
+            $CORSUrls += 'https://functions-staging.azure.com'
+            $CORSUrls += 'https://functions-next.azure.com'
+        }
+    }
+
+    Write-Output $CORSUrls
+
+    Write-Footer -ScopedPSCmdlet $PSCmdlet
+}
+
+function Set-CorsSettings
+{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)][string] [ValidateSet('functionapp', 'webapp')]$AppType,
+        [Parameter(Mandatory)][string[]] $CORSUrls,
+        [Parameter(Mandatory)][string] $ResourceId,
+        [Parameter()][string] $ResourceDeploymentSlotName
+    )
+    
+    Write-Header -ScopedPSCmdlet $PSCmdlet
+
+    $optionalParameters = @()
+    if ($ResourceDeploymentSlotName)
+    {
+        $optionalParameters += '--slot', "$ResourceDeploymentSlotName"
+    }
+
+    $currentCorsSettings = Invoke-Executable az $AppType cors show --ids $ResourceId @optionalParameters | ConvertFrom-Json
+
+    [string[]]$currentCorsOrigins = @()
+    if ($currentCorsSettings -and $currentCorsSettings.allowedOrigins)
+    {
+        $currentCorsOrigins = $currentCorsSettings.allowedOrigins
+    }
+
+    Compare-Object -ReferenceObject $currentCorsOrigins -DifferenceObject $CORSUrls | ForEach-Object {
+        $value = $_.InputObject
+        $sideIndicator = $_.SideIndicator
+        switch ($sideIndicator)
+        {
+            '=>'
+            { 
+                Write-Host "Adding CORS URL $value"
+                Invoke-Executable az $AppType cors add --ids $ResourceId --allowed-origins $value @optionalParameters
+            }
+            '<='
+            {
+                Write-Host "Removing CORS URL $value"
+                Invoke-Executable az $AppType cors remove --ids $ResourceId --allowed-origins $value @optionalParameters
+            }
         }
     }