+++ src/AzDocs.Common/public/Rest-Helper-Function.ps1
@@ -7,10 +7,9 @@ function Invoke-AzRestCall
         [Parameter()][string] $ResourceId,
         [Parameter()][string] $ResourceUrl,
         [Parameter(Mandatory)][string] $ApiVersion, # Example: "2021-02-01-preview"
-        [Parameter(Mandatory)][PSCustomObject] $Body
+        [Parameter()][PSCustomObject] $Body,
+        [Parameter()][switch] $AllowToFail
     )
-
-    $json = ($Body | ConvertTo-Json -Compress -Depth 100).Replace("""", """""")
 
     if (!$ResourceUrl -and !$ResourceId)
     {
@@ -27,6 +26,32 @@ function Invoke-AzRestCall
     {
         $url = "$($ResourceId)?api-version=$($ApiVersion)"
     }
+    
+    if (!$Body)
+    {
+        Invoke-Executable -AllowToFail:$AllowToFail az rest --method $Method --url $url 
+    }
+    else
+    {
+        $json = ($Body | ConvertTo-Json -Compress -Depth 100).Replace("""", """""")
+        Invoke-Executable -AllowToFail:$AllowToFail az rest --method $Method --url $url --body """$json"""
+    }
+}
 
-    Invoke-Executable az rest --method $Method --url $url --body """$json"""
+function Show-RestError
+{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)][PSObject] $Exception
+    )
+
+    Write-Host 'Response: ' $Exception -ForegroundColor Red
+    Write-Host 'StatusCode:' $Exception.Exception.Response.StatusCode.value__ -ForegroundColor Red
+    Write-Host 'Reason:' $Exception.Exception.Response.ReasonPhrase -ForegroundColor Red
+    if ($Exception.Exception.Response.StatusDescription)
+    {
+        Write-Host 'StatusDescription:' $Exception.Exception.Response.StatusDescription -ForegroundColor Red
+    }
+    Write-Host 'Exception:' $Exception.Exception -ForegroundColor Red
+    throw $Exception;
 }